Fixing Drone dynamics, Setting up drone stop

+++ include/ui_handler.h
@@ -4,7 +4,7 @@
 #include <ncurses.h>
 
 // Display the status window with the drone's position and state
-void display_status(int x, int y, const char *status);
+void display_status(float x, float y, const char *status);
 
 // Display a key legend at the bottom of the screen
 void display_legend();
